+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Research Project updates [#2540](https://github.com/rokwire/illinois-app/issues/2540).
+
 ## [4.2.20] - 2022-11-16
 ### Changed
 - Research Project updates [#2533](https://github.com/rokwire/illinois-app/issues/2533).